🐞 fix: use abs path for docker
