+++ app/api/core/converters/doc/doc_converter.py
@@ -1,8 +1,9 @@
-from ..base_converter import (
+from api.core.converters.base_converter import (
     BaseConverter,
 )
 
 
-class PdfConverter(BaseConverter):
+class DocConverter(BaseConverter):
+
     def convert(self, *args, **kwargs) -> str:
         raise NotImplementedError("Subclasses must implement this method")
