+++ app/api/core/converters/doc/epub_converter.py
@@ -1,8 +1,8 @@
-from ..base_converter import (
+from api.core.converters.base_converter import (
     BaseConverter,
 )
 
 
-class HtmlConverter(BaseConverter):
+class EpubConverter(BaseConverter):
     def convert(self, *args, **kwargs) -> str:
         raise NotImplementedError("Subclasses must implement this method")
