+++ app/api/core/converters/doc/html_converter.py
@@ -1,9 +1,8 @@
-from ..base_converter import (
+from api.core.converters.base_converter import (
     BaseConverter,
 )
 
 
-class DocConverter(BaseConverter):
-
+class HtmlConverter(BaseConverter):
     def convert(self, *args, **kwargs) -> str:
         raise NotImplementedError("Subclasses must implement this method")
